ledstrip: add presets
this commit adds 4 presets to the ledstrip StatusModeConfig.

For the actual LED strip control, first check if any preset is actived
by the RC MDOE control.
If it is not, use the configured ledstrip_preset_default.

In the CLI, the 4 ledstrip presets will be accessible by an optional
"preset N" argument, defaulting to preset 0.

TODO: in the MSP, have the possibility to change the
ledstrip_preset_default.

Signed-off-by: Tim Froidcoeur <tim.froidcoeur@gmail.com>